Add better ngrok waiting and multiple agents
To better assist in making sure the ngrok service is up and running, and
has a tunnel established. The `medici-startup.sh` has a wait loop that
will attempt to connect to ngrok and wait till it has a valid request
before proceeding

There is also a second docker-compose file that will fire up multiple
agents for testing propses.

The ngrok utility also has logging enabled to assist in making sure the
ngrok service is alive and healthy.